🚚 refactor: move views
